feat: optimise store date planets
