+++ api/collector/collector.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 	"os"
 
+	"github.com/Spacelocust/for-democracy/collector/hellhub"
 	"github.com/bytedance/sonic"
 )
 
@@ -20,30 +21,6 @@ type DataHellhub[T any] struct {
 type Fetch struct {
 	Url        string
 	PathPrefix string
-}
-
-func hellhubFetch[T any](url string) ([]T, error) {
-	var data struct {
-		List []T `json:"data"`
-	}
-
-	resp, err := http.Get(fmt.Sprintf("%s%s%s", os.Getenv("HELLHUB_API_URL"), "/api", url))
-	if err != nil {
-		return nil, fmt.Errorf("failed to make HTTP request: %w", err)
-	}
-
-	defer resp.Body.Close()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read response body: %w", err)
-	}
-
-	if err := sonic.Unmarshal(body, &data); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal response body: %w", err)
-	}
-
-	return data.List, nil
 }
 
 func helldiversFetch[T any](url string, prefix ...bool) ([]T, error) {
@@ -74,19 +51,15 @@ func helldiversFetch[T any](url string, prefix ...bool) ([]T, error) {
 }
 
 func GatherData() {
-	// if err := getStratagems(); err != nil {
+	if err := hellhub.GetData(); err != nil {
+		fmt.Println(err)
+	}
+
+	// if err := GetDefences(); err != nil {
 	// 	fmt.Println(err)
 	// }
 
-	if err := getPlanets(); err != nil {
-		fmt.Println(err)
-	}
-
-	if err := GetDefences(); err != nil {
-		fmt.Println(err)
-	}
-
-	if err := GetLiberations(); err != nil {
-		fmt.Println(err)
-	}
+	// if err := GetLiberations(); err != nil {
+	// 	fmt.Println(err)
+	// }
 }
